requirements.txt
openCV-python
pillow
numpy
moviepy